LayoutTests/imported/w3c/web-platform-tests/content-security-policy/style-src-attr-elem/style-src-attr-blocked-src-allowed-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'style-src-attr'.


Harness Error (TIMEOUT), message = null

NOTRUN Should fire a security policy violation event
FAIL The attribute style should not be applied assert_equals: expected 0 but got 10
